+++ packages/fluentui/docs/src/examples/components/Avatar/index.tsx
@@ -5,6 +5,7 @@ import Types from './Types';
 import Variations from './Variations';
 import Usage from './Usage';
 import Performance from './Performance';
+import Visual from './Visual';
 
 const AvatarExamples = () => (
   <>
@@ -12,6 +13,7 @@ const AvatarExamples = () => (
     <Variations />
     <Rtl />
     <Usage />
+    <Visual />
     <Performance />
   </>
 );