+++ packages/react-internal/src/components/SpinButton/SpinButton.base.tsx
@@ -14,13 +14,15 @@ import {
 import { getArrowButtonStyles } from './SpinButton.styles';
 import { ISpinButtonProps, ISpinButtonStyleProps, ISpinButtonStyles, KeyboardSpinDirection } from './SpinButton.types';
 import { Position } from '../../Positioning';
-import { useSetTimeout, useControllableValue, useWarnings, useId } from '@fluentui/react-hooks';
+import { useAsync, useControllableValue, useWarnings, useId, usePrevious } from '@fluentui/react-hooks';
 
 interface ISpinButtonInternalState {
-  lastValidValue?: string;
   spinningByMouse?: boolean;
-  valueToValidate?: string;
   stepTimeoutHandle: number;
+  /** Allows access to the latest `value` inside reused callbacks (to avoid stale capture issues) */
+  latestValue: string | undefined;
+  /** Allows access to the latest `intermediateValue` inside reused callbacks (to avoid stale capture issues) */
+  latestIntermediateValue: string | undefined;
 }
 
 const getClassNames = classNamesFunction<ISpinButtonStyleProps, ISpinButtonStyles>();
@@ -28,7 +30,7 @@ const getClassNames = classNamesFunction<ISpinButtonStyleProps, ISpinButtonStyle
 const COMPONENT_NAME = 'SpinButton';
 const DEFAULT_PROPS: Required<Pick<
   ISpinButtonProps,
-  // These are explicitly specified so that the only the things which actually have defaults
+  // These are explicitly specified so that only the things which actually have defaults
   // get marked as required in ISpinButtonPropsWithDefaults below
   'disabled' | 'label' | 'step' | 'labelPosition' | 'incrementButtonIcon' | 'decrementButtonIcon'
 >> = {
@@ -65,7 +67,7 @@ const useComponentRef = (
   );
 };
 
-const onChange = (): void => {
+const noOp = (): void => {
   /**
    * A noop input change handler. Using onInput instead of onChange was meant to address an issue
    * which apparently has been resolved in React 16 (https://github.com/facebook/react/issues/7027).
@@ -75,6 +77,17 @@ const onChange = (): void => {
    * TODO (Fabric 8?) - switch to just calling onChange (this is a breaking change for any tests,
    * ours or 3rd-party, which simulate entering text in a SpinButton)
    */
+};
+
+/** Clamp the value to the provided min and/or max */
+const clampValue = (value: number, { min, max }: { max?: number; min?: number }) => {
+  if (typeof max === 'number') {
+    value = Math.min(value, max);
+  }
+  if (typeof min === 'number') {
+    value = Math.max(value, min);
+  }
+  return value;
 };
 
 export const SpinButtonBase: React.FunctionComponent<ISpinButtonProps> = React.forwardRef<
@@ -112,6 +125,7 @@ export const SpinButtonBase: React.FunctionComponent<ISpinButtonProps> = React.f
     onIncrement,
     iconButtonProps,
     onValidate,
+    onChange,
     styles,
   } = props;
 
@@ -121,23 +135,43 @@ export const SpinButtonBase: React.FunctionComponent<ISpinButtonProps> = React.f
 
   const [isFocused, setIsFocused] = React.useState(false);
   const [keyboardSpinDirection, setKeyboardSpinDirection] = React.useState(KeyboardSpinDirection.notSpinning);
-  const { setTimeout, clearTimeout } = useSetTimeout();
-
-  const [value, setValue] = useControllableValue(valueFromProps, defaultValue ?? String(min || 0));
-
-  const { current: internalState } = React.useRef<ISpinButtonInternalState>({
-    lastValidValue: value,
-    stepTimeoutHandle: -1,
-  });
-
-  if (typeof valueFromProps === 'string') {
-    // Ensure that we respect updates to props.value when determining the last valid value
-    internalState.lastValidValue = valueFromProps;
-  }
+  const async = useAsync();
 
   const precision = React.useMemo(() => {
     return precisionFromProps ?? Math.max(calculatePrecision(step), 0);
   }, [precisionFromProps, step]);
+
+  /**
+   * Actual current value. If `props.value` is provided (controlled), it will always be used.
+   * If not (uncontrolled), this tracks the current value based on user modifications.
+   * Note that while the user is editing text in the field, this will not be updated until "commit"
+   * (blur or press enter).
+   */
+  const [value, setValue] = useControllableValue(valueFromProps, defaultValue ?? String(min || 0), onChange);
+  /**
+   * "Uncommitted" internal value while the user is editing text in the field. This lets us wait to
+   * call `onChange` (and possibly update the real value) until the user "commits" the value by
+   * pressing enter or blurring the field.
+   */
+  const [intermediateValue, setIntermediateValue] = React.useState<string>();
+
+  const { current: internalState } = React.useRef<ISpinButtonInternalState>({
+    stepTimeoutHandle: -1,
+    latestValue: undefined,
+    latestIntermediateValue: undefined,
+  });
+  // On each render, update this saved value used by callbacks. (This should be safe even if render
+  // is called multiple times, because an event handler or timeout callback will only run once.)
+  internalState.latestValue = value;
+  internalState.latestIntermediateValue = intermediateValue;
+
+  const previousValueFromProps = usePrevious(valueFromProps);
+  React.useEffect(() => {
+    // If props.value changes while editing, clear the intermediate value
+    if (valueFromProps !== previousValueFromProps && intermediateValue !== undefined) {
+      setIntermediateValue(undefined);
+    }
+  }, [valueFromProps, previousValueFromProps, intermediateValue]);
 
   const classNames = getClassNames(styles, {
     theme: theme!,
@@ -154,103 +188,99 @@ export const SpinButtonBase: React.FunctionComponent<ISpinButtonProps> = React.f
     'className',
   ]);
 
-  /** Composed validation handler (uses `props.onValidate` or default) */
-  const handleValidate = (newValue: string, ev: React.SyntheticEvent<HTMLInputElement>) => {
-    if (onValidate) {
-      return onValidate(newValue, ev);
-    }
-    // default validation handling
-    let newNumber = Number(newValue);
-    if (newValue && newValue.trim().length && !isNaN(newNumber)) {
-      if (typeof min === 'number') {
-        newNumber = Math.max(min, newNumber);
+  /** Validate (commit) function called on blur or enter keypress. */
+  const validate = React.useCallback(
+    (ev: React.SyntheticEvent<HTMLElement>): void => {
+      // Only run validation if the value changed
+      const enteredValue = internalState.latestIntermediateValue;
+      if (enteredValue !== undefined && enteredValue !== internalState.latestValue) {
+        let newValue: string | undefined;
+        if (onValidate) {
+          newValue = onValidate(enteredValue, ev) as string | undefined;
+        } else if (enteredValue && enteredValue.trim().length && !isNaN(Number(enteredValue))) {
+          // default validation handling
+          newValue = String(clampValue(Number(enteredValue), { min, max }));
+        }
+        if (newValue !== undefined && newValue !== internalState.latestValue) {
+          // Commit the value if it changed
+          setValue(newValue);
+        }
       }
-      if (typeof max === 'number') {
-        newNumber = Math.min(max, newNumber);
-      }
-      return String(newNumber);
-    }
-  };
 
-  /** Validate function called on blur or enter keypress. */
-  const validate = (ev: React.FocusEvent<HTMLInputElement> | React.KeyboardEvent<HTMLInputElement>): void => {
-    if (
-      value !== undefined &&
-      internalState.valueToValidate !== undefined &&
-      internalState.valueToValidate !== internalState.lastValidValue
-    ) {
-      const newValue = handleValidate(internalState.valueToValidate, ev);
-
-      // Done validating this value, so clear it
-      internalState.valueToValidate = undefined;
-      if (newValue !== undefined) {
-        internalState.lastValidValue = newValue;
-        setValue(newValue);
-      } else {
-        // Value was invalid. Reset state to last valid value.
-        setValue(internalState.lastValidValue);
-      }
-    }
-  };
+      // Done validating, so clear the intermediate typed value (if any)
+      setIntermediateValue(undefined);
+    },
+    [internalState, max, min, onValidate, setValue],
+  );
 
   /**
    * Stop spinning (clear any currently pending update and set spinning to false)
    */
   const stop = React.useCallback((): void => {
     if (internalState.stepTimeoutHandle >= 0) {
-      clearTimeout(internalState.stepTimeoutHandle);
+      async.clearTimeout(internalState.stepTimeoutHandle);
       internalState.stepTimeoutHandle = -1;
     }
     if (internalState.spinningByMouse || keyboardSpinDirection !== KeyboardSpinDirection.notSpinning) {
       internalState.spinningByMouse = false;
       setKeyboardSpinDirection(KeyboardSpinDirection.notSpinning);
     }
-  }, [internalState, keyboardSpinDirection, clearTimeout]);
+  }, [internalState, keyboardSpinDirection, async]);
 
   /**
-   * Update the value with the given stepFunction
+   * Update the value with the given stepFunction.
+   * Also starts spinning for mousedown events by scheduling another update with setTimeout.
    * @param stepFunction - function to use to step by
    * @param event - The event that triggered the updateValue
-   * @param shouldSpin - should we fire off another updateValue when we are done here? This should be true
-   * when spinning in response to a mousedown. It doesn't need to be true for keyboard spinning since
-   * holding the key will automatically fire more keydown events.
    */
   const updateValue = React.useCallback(
     (
-      stepFunction: (
-        value: string,
-        event?: React.MouseEvent<HTMLElement> | React.KeyboardEvent<HTMLElement>,
-      ) => string | void,
-      ev?: React.MouseEvent<HTMLElement> | React.KeyboardEvent<HTMLElement>,
-      shouldSpin: boolean = false,
+      stepFunction: Required<ISpinButtonProps>['onIncrement'],
+      ev: React.MouseEvent<HTMLElement> | React.KeyboardEvent<HTMLInputElement>,
     ): void => {
-      setValue(prevValue => {
-        // For spinning by mouse, where additional updateValue calls are triggered via setTimeout,
-        // we must access the current value in an updater callback (rather than referencing `value`
-        // directly) because otherwise we'd be reusing the stale captured value from the first call
-        // and spinning wouldn't work. (The other possible approach here is storing the value in
-        // internalState while spinning, then setting the actual state when spinning stops.)
-        const newValue = stepFunction(prevValue || '', ev);
-        if (newValue !== undefined) {
-          internalState.lastValidValue = newValue;
-          return newValue;
+      ev.persist();
+
+      if (internalState.latestIntermediateValue !== undefined) {
+        // Edge case: if intermediateValue is set, this means that the user was editing the input
+        // text and then started spinning (either with mouse or keyboard). We need to validate and
+        // call onChange before starting to spin.
+        if (ev.type === 'keydown') {
+          // For the arrow keys, we have to manually trigger validation.
+          // (For the buttons, validation will happen automatically since the input's onBlur will
+          // be triggered after mousedown on the button completes.)
+          validate(ev);
         }
-        return prevValue;
-      });
+        async.requestAnimationFrame(() => {
+          // After handling any value updates, do the spinning update
+          updateValue(stepFunction, ev);
+        });
+        return;
+      }
 
+      // Call the step function and update the value.
+      // (Note: we access the latest value via internalState (not directly) to ensure we don't use
+      // a stale captured value. This is mainly important for spinning by mouse, where we trigger
+      // additional calls to the original updateValue function via setTimeout. It also lets us
+      // avoid useCallback deps on frequently changing values.)
+      const newValue = stepFunction(internalState.latestValue || '', ev) as string | undefined;
+      if (newValue !== undefined && newValue !== internalState.latestValue) {
+        setValue(newValue);
+      }
+
+      // Schedule the next spin if applicable
+      // (will be canceled if there's a mouseup before the timeout runs)
       const wasSpinning = internalState.spinningByMouse;
-      internalState.spinningByMouse = shouldSpin;
-
-      if (shouldSpin) {
-        internalState.stepTimeoutHandle = setTimeout(
+      internalState.spinningByMouse = ev.type === 'mousedown';
+      if (internalState.spinningByMouse) {
+        internalState.stepTimeoutHandle = async.setTimeout(
           () => {
-            updateValue(stepFunction, ev, true);
+            updateValue(stepFunction, ev);
           },
-          wasSpinning ? STEP_DELAY : INITIAL_STEP_DELAY,
+          wasSpinning ? STEP_DELAY : INITIAL_STEP_DELAY, // the first step is slower
         );
       }
     },
-    [internalState, setValue, setTimeout],
+    [internalState, async, validate, setValue],
   );
 
   /** Composed increment handler (uses `props.onIncrement` or default) */
@@ -259,10 +289,7 @@ export const SpinButtonBase: React.FunctionComponent<ISpinButtonProps> = React.f
       if (onIncrement) {
         return onIncrement(newValue);
       } else {
-        let numericValue = Number(newValue) + Number(step);
-        if (typeof max === 'number') {
-          numericValue = Math.min(numericValue, max);
-        }
+        let numericValue = clampValue(Number(newValue) + Number(step), { max });
         numericValue = precisionRound(numericValue, precision);
         return String(numericValue);
       }
@@ -276,10 +303,7 @@ export const SpinButtonBase: React.FunctionComponent<ISpinButtonProps> = React.f
       if (onDecrement) {
         return onDecrement(newValue);
       } else {
-        let numericValue = Number(newValue) - Number(step);
-        if (typeof min === 'number') {
-          numericValue = Math.max(numericValue, min);
-        }
+        let numericValue = clampValue(Number(newValue) - Number(step), { min });
         numericValue = precisionRound(numericValue, precision);
         return String(numericValue);
       }
@@ -288,10 +312,8 @@ export const SpinButtonBase: React.FunctionComponent<ISpinButtonProps> = React.f
   );
 
   /** Handles when the user types in the input */
-  const handleInputChange = (ev: React.FormEvent<HTMLInputElement | HTMLTextAreaElement>): void => {
-    const elementValue = (ev.target as HTMLInputElement).value;
-    internalState.valueToValidate = elementValue;
-    setValue(elementValue);
+  const handleInputChange = (ev: React.FormEvent<HTMLInputElement>): void => {
+    setIntermediateValue((ev.target as HTMLInputElement).value);
   };
 
   /** Composed focus handler (does internal stuff and calls `props.onFocus`) */
@@ -340,10 +362,12 @@ export const SpinButtonBase: React.FunctionComponent<ISpinButtonProps> = React.f
         updateValue(handleDecrement, ev);
         break;
       case KeyCodes.enter:
+        // Commit the edited value
         validate(ev);
         break;
       case KeyCodes.escape:
-        setValue(internalState.lastValidValue);
+        // Revert to previous value
+        setIntermediateValue(undefined);
         break;
     }
     // style the increment/decrement button to look active
@@ -366,20 +390,22 @@ export const SpinButtonBase: React.FunctionComponent<ISpinButtonProps> = React.f
 
   const handleIncrementMouseDown = React.useCallback(
     (ev: React.MouseEvent<HTMLElement>): void => {
-      updateValue(handleIncrement, ev, true /* shouldSpin */);
+      updateValue(handleIncrement, ev);
     },
     [handleIncrement, updateValue],
   );
 
   const handleDecrementMouseDown = React.useCallback(
     (ev: React.MouseEvent<HTMLElement>): void => {
-      updateValue(handleDecrement, ev, true /* shouldSpin */);
+      updateValue(handleDecrement, ev);
     },
     [handleDecrement, updateValue],
   );
 
   useComponentRef(props, input, value);
   useDebugWarnings(props);
+
+  const valueIsNumber = !!value && !isNaN(Number(value)); // Number('') is 0 which may not be desirable
 
   const labelContent = (iconProps || label) && (
     <div className={classNames.labelWrapper}>
@@ -404,23 +430,19 @@ export const SpinButtonBase: React.FunctionComponent<ISpinButtonProps> = React.f
         data-ktp-target={true}
       >
         <input
-          value={value}
+          // Display intermediateValue while editing the text (before commit)
+          value={intermediateValue ?? value}
           id={inputId}
-          onChange={onChange}
+          onChange={noOp}
           onInput={handleInputChange}
           className={classNames.input}
           type="text"
           autoComplete="off"
           role="spinbutton"
           aria-labelledby={label && labelId}
-          aria-valuenow={
-            typeof ariaValueNow === 'number'
-              ? ariaValueNow
-              : value && !isNaN(Number(value)) // Number('') is 0 which may not be desirable
-              ? Number(value)
-              : undefined
-          }
-          aria-valuetext={ariaValueText ? ariaValueText : !value || isNaN(Number(value)) ? value : undefined}
+          // TODO: test what happens while editing
+          aria-valuenow={ariaValueNow ?? (valueIsNumber ? Number(value) : undefined)}
+          aria-valuetext={ariaValueText ?? (valueIsNumber ? undefined : value)}
           aria-valuemin={min}
           aria-valuemax={max}
           aria-describedby={ariaDescribedBy}