+++ packages/react-internal/src/components/TextField/TextField.base.tsx
@@ -243,7 +243,8 @@ export class TextFieldBase extends React.Component<ITextFieldProps, ITextFieldSt
             {multiline ? this._renderTextArea() : this._renderInput()}
             {iconProps && <Icon className={classNames.icon} {...iconProps} />}
             {hasRevealButton && (
-              <button className={classNames.revealButton} onClick={this._onRevealButtonClick}>
+              // Explicitly set type="button" since the default button type within a form is "submit"
+              <button className={classNames.revealButton} onClick={this._onRevealButtonClick} type="button">
                 <span className={classNames.revealSpan}>
                   <Icon
                     className={classNames.revealIcon}